+++ Apps/Playground/index.js
@@ -3,7 +3,7 @@
  */
 
 import {AppRegistry} from 'react-native';
-import App from './App';
-import {name as appName} from './app.json';
+import App from './shared/App';
+import {name as appName} from './shared/app.json';
 
 AppRegistry.registerComponent(appName, () => App);
